[JaKö]  end-to-end implementation for US1.1A.1
Frontend has a working form for a reservation as a guest but no error or notifacation handling.

Backend catches the request and saves a reservation in the DB. A new guest user is also saved in the DB.
It then returns the newly created reservation.

No authentication for the ReservationEndpoint (currently using the unsecure path "authentication" which is why the LoginEndpoint is disabled).

Place (table) is never assigned in current implementation.

User entity was deleted (as well as some classes refering to it). Instead, ApplicationUser now handles all of the User data.